rename require-auth component file to Auth
